refactor: Properties로 네이밍 수정
