+++ src/main/java/kr/pickple/back/auth/config/http/KakaoMemberHttpInterfaceConfig.java
@@ -1,4 +1,4 @@
-package kr.pickple.back.auth.config;
+package kr.pickple.back.auth.config.http;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,7 +6,8 @@ import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.support.WebClientAdapter;
 import org.springframework.web.service.invoker.HttpServiceProxyFactory;
 
-import kr.pickple.back.auth.config.property.KakaoOAuthConfig;
+import kr.pickple.back.auth.config.WebClientConfig;
+import kr.pickple.back.auth.config.property.KakaoOAuthProperties;
 import kr.pickple.back.auth.service.memberclient.KakaoMemberApiClient;
 import lombok.RequiredArgsConstructor;
 
@@ -14,8 +15,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class KakaoMemberHttpInterfaceConfig {
 
-    private final KakaoOAuthConfig kakaoOauthConfig;
-    private final WebClientConfig webClientConfig;
+    private final KakaoOAuthProperties kakaoOauthProperties;
 
     @Bean
     public KakaoMemberApiClient kakaoMemberApiClient() {
@@ -24,8 +24,8 @@ public class KakaoMemberHttpInterfaceConfig {
 
     private <T> T createHttpInterface(final Class<T> clazz) {
         final WebClient webClient = WebClient.builder()
-                .baseUrl(kakaoOauthConfig.getMemberUrl())
-                .exchangeStrategies(webClientConfig.getExchangeStrategies())
+                .baseUrl(kakaoOauthProperties.getMemberUrl())
+                .exchangeStrategies(WebClientConfig.getExchangeStrategies())
                 .build();
         final HttpServiceProxyFactory build = HttpServiceProxyFactory
                 .builder(WebClientAdapter.forClient(webClient))
