+++ src/main/java/kr/pickple/back/common/config/WebConfig.java
@@ -6,7 +6,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import kr.pickple.back.auth.config.property.CorsConfig;
+import kr.pickple.back.auth.config.property.CorsProperties;
 import kr.pickple.back.auth.util.OAuthProviderConverter;
 import lombok.RequiredArgsConstructor;
 
@@ -14,13 +14,13 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
-    private final CorsConfig corsConfig;
+    private final CorsProperties corsProperties;
 
     @Override
     public void addCorsMappings(final CorsRegistry registry) {
 
         registry.addMapping("/**")
-                .allowedOrigins(corsConfig.getUrl())
+                .allowedOrigins(corsProperties.getUrl())
                 .allowedMethods(
                         HttpMethod.GET.name(),
                         HttpMethod.POST.name(),
